串口增加对空闲中断的完全使能，可同时使用两种模式

+++ Device/Inc/dev_serial.h
@@ -22,7 +22,7 @@ class Serialctrl : public Serialdev
 {
 public:
 	Serialctrl(USART_TypeDef *_USARTx, uint32_t BufferSize) : Serialdev(_USARTx, BufferSize) {}
-	Serialctrl(USART_TypeDef *_USARTx, uint32_t BufferSize, uint16_t USART_ITPending) : Serialdev(_USARTx, BufferSize, USART_ITPending) {}
+	Serialctrl(USART_TypeDef *_USARTx, uint32_t BufferSize, uint32_t USART_ITPending) : Serialdev(_USARTx, BufferSize, USART_ITPending) {}
 
 	void attachInterrupt(USART_CallbackFunction_t Function);
 	void IRQHandler(void);
